Move votes and delegate tab to the wallet directory
