+++ src/components/screens/wallet/explorer.js
@@ -9,8 +9,8 @@ import { getAccount } from '../../../utils/api/account';
 import { getTransactions } from '../../../utils/api/transactions';
 import txFilters from '../../../constants/transactionFilters';
 import TabsContainer from '../../toolbox/tabsContainer/tabsContainer';
-import DelegateTab from '../../shared/delegate';
-import VotesTab from '../../shared/votes';
+import DelegateTab from './delegateProfile';
+import VotesTab from './votes';
 import Transactions from './transactions';
