+++ src/components/screens/wallet/index.js
@@ -7,8 +7,8 @@ import Overview from './overview';
 import { getTransactions } from '../../../actions/transactions';
 import txFilters from '../../../constants/transactionFilters';
 import TabsContainer from '../../toolbox/tabsContainer/tabsContainer';
-import DelegateTab from '../../shared/delegate';
-import VotesTab from '../../shared/votes';
+import DelegateTab from './delegateProfile';
+import VotesTab from './votes';
 import Transactions from './transactions';
 import { isEmpty } from '../../../utils/helpers';
 import Send from '../send';
