+++ src/components/screens/wallet/votes/votes.css
@@ -1,4 +1,4 @@
-@import '../../../app/mixins.css';
+@import '../../../../app/mixins.css';
 
 .wrapper {
   margin-bottom: 40px;
